+++ mangadex_library/lib/user/user_followed_groups/group_check.dart
@@ -1,0 +1,13 @@
+class GroupCheck {
+	String? result;
+
+	GroupCheck({this.result});
+
+	factory GroupCheck.fromJson(Map<String, dynamic> json) => GroupCheck(
+				result: json['result'] as String?,
+			);
+
+	Map<String, dynamic> toJson() => {
+				'result': result,
+			};
+}