+++ mangadex_library/lib/user/user_followed_manga/manga_check.dart
@@ -1,0 +1,13 @@
+class MangaCheck {
+	String? result;
+
+	MangaCheck({this.result});
+
+	factory MangaCheck.fromJson(Map<String, dynamic> json) => MangaCheck(
+				result: json['result'] as String?,
+			);
+
+	Map<String, dynamic> toJson() => {
+				'result': result,
+			};
+}