+++ mangadex_library/lib/user/user_followed_users/user_check.dart
@@ -1,0 +1,13 @@
+class UserCheck {
+	String? result;
+
+	UserCheck({this.result});
+
+	factory UserCheck.fromJson(Map<String, dynamic> json) => UserCheck(
+				result: json['result'] as String?,
+			);
+
+	Map<String, dynamic> toJson() => {
+				'result': result,
+			};
+}